fix: P1/P2 correctness — memcpy truncation, stack pass compat, path diagnostics
- createMemcpy: loop counter int -> uint64_t (signed/unsigned mismatch
  with uint64_t C_size; truncation on copies > INT_MAX)
- ResizeAllocatedStackPass: skip i8 allocas produced by the rewritten
  PromotePseudoStackPass (the /16 divisor and chainLook logic assumed
  the legacy i128 format)
- lift_jmp/lift_ret: capture solvePath return, increment
  blocks_unreachable and emit [diag] line on PATH_unsolved

+++ lifter/memory/GEPTracker.ipp
@@ -57,7 +57,7 @@ MERGEN_LIFTER_DEFINITION_TEMPLATES(void)::createMemcpy(Value* src, Value* dest,
     return;
   }
 
-  for (int i = 0; i < C_size; i++) {
+  for (uint64_t i = 0; i < C_size; i++) {
     if (!buffer.contains(C_src + i)) {
       printvalue2(C_src + i);
       printvalue2(C_dest + i);
